userId long 으로 변환

+++ BookieProject/src/main/java/com/project/bookie/mapper/user/UserAuthMapper.java
@@ -19,12 +19,12 @@ public interface UserAuthMapper {
 	public void updateUserWithKey(@Param("uId") long uId, @Param("authKey") String authKey);
 
 	// 해당 아이디가 이미 인증된 이메일인지 확인
-	public int checkCertification(int uId);
+	public int checkCertification(long uId);
 
 	// 해당 아이디로 존재하는 키가 있는지 확인
-	public String selectKeyWithId(int uId);
+	public String selectKeyWithId(long uId);
 	// selectKeyWithId
 
 	// 키가 있을 경우 해당 아이디의 인증 여부를 변경
-	public void updateUserCertification(int uId);
+	public void updateUserCertification(long uId);
 }
